requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

langchain>=0.2.10
langchain-community>=0.2.9
langchain-groq>=0.1.6

sentence-transformers>=2.6.1
faiss-cpu>=1.8.0

pydantic>=2.6.0
python-dotenv>=1.0.1

unstructured>=0.12.6
pdfminer.six>=20231228
pypdf>=4.2.0

numpy>=1.26.0
